feat(component): move time localization logic to relevant file
