+++ packages/big-design/src/utils/localizationProvider/localizationProvider.ts
@@ -1,3 +1,16 @@
+function getTimeIntervals24hr() {
+  const times = ['00:00'];
+
+  for (let i = 1; i < 24; i++) {
+    times.push(`${i}:00`);
+  }
+  times.push('23:59');
+
+  return times.map((time) => ({ value: time, content: time }));
+}
+
+const defaultTimeIntervals = getTimeIntervals24hr();
+
 export const createLocalizationProvider = (locale: string) => {
   const dayFormatter = Intl.DateTimeFormat(locale, {
     weekday: 'short',
@@ -29,3 +42,20 @@ export const createLocalizationProvider = (locale: string) => {
     formatTime: timeFormatter.format,
   };
 };
+
+export function getTimeIntervals(locale: string) {
+  const localization = createLocalizationProvider(locale);
+  const localizedTimeIntervals = defaultTimeIntervals.map((time) => {
+    const baseDate = new Date();
+    const [hour, minute] = time.value.split(':');
+    baseDate.setHours(parseInt(hour, 10));
+    baseDate.setMinutes(parseInt(minute, 0));
+
+    return {
+      content: localization.formatTime(baseDate),
+      value: time.value,
+    };
+  });
+
+  return localizedTimeIntervals;
+}
